# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Safalife API is a comprehensive backend service for Safalife's competition management system. This API provides endpoints for health monitoring, user management, competition tracking, and administrative functions. Built with Go using Clean Architecture principles, it offers high performance and reliability for managing large-scale competitions.",
        "title": "Safalife API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Safalife API Support Team",
            "url": "http://www.safalife.com/support",
            "email": "api-support@safalife.com"
        },
        "license": {
            "name": "MIT License",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8080",
    "basePath": "/v1",
    "paths": {
...
        "/quran/bookmarks/ayahs": {
            "post": {
                "description": "Bookmark an ayah for the current user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bookmarks"
                ],
                "summary": "Create ayah bookmark",
                "parameters": [
                    {
                        "description": "Create bookmark request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateBookmarkAyahRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Bookmark created successfully",
                        "schema": {
                            "$ref": "#/definitions/quran.BookmarkAyahSuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/quran.QuranErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/quran.QuranErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/quran/bookmarks/ayahs/{id}": {
            "get": {
                "description": "Get a specific bookmark by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bookmarks"
                ],
                "summary": "Get bookmark ayah by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bookmark ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "enum": [
                            "ayah"
                        ],
                        "type": "string",
                        "example": "ayah",
                        "description": "Include related data",
                        "name": "include",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Bookmark retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/quran.BookmarkAyahSuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/quran.QuranErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Bookmark not found",
                        "schema": {
                            "$ref": "#/definitions/quran.QuranErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/quran.QuranErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a bookmark by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bookmarks"
                ],
                "summary": "Delete bookmark ayah",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bookmark ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Bookmark deleted successfully",
                        "schema": {
                            "$ref": "#/definitions/quran.DeleteBookmarkAyahSuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/quran.QuranErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Bookmark not found",
                        "schema": {
                            "$ref": "#/definitions/quran.QuranErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/quran.QuranErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/quran/last-reads": {
            "post": {
                "description": "Create a new last read entry for a user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "LastRead"
                ],
                "summary": "Create last read",
                "parameters": [
                    {
                        "description": "LastRead information",
                        "name": "lastRead",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateLastReadRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Last read created successfully",
                        "schema": {
                            "$ref": "#/definitions/quran.LastReadSuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/quran.QuranErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/quran.QuranErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/quran/progress-hatam": {
            "post": {
                "description": "Create a new progress hatam entry for a user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ProgressHatam"
                ],
                "summary": "Create progress hatam",
                "parameters": [
                    {
                        "description": "ProgressHatam information",
                        "name": "progressHatam",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateProgressHatamRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Progress hatam created successfully",
                        "schema": {
                            "$ref": "#/definitions/quran.ProgressHatamSuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/quran.QuranErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/quran.QuranErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/quran/surahs": {
            "get": {
                "description": "Get all surahs from the Quran with pagination",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Surahs"
                ],
                "summary": "Get all surahs",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Limit (0 for all)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Offset",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "ayahs"
                        ],
                        "type": "string",
                        "example": "ayahs",
                        "description": "Include related data",
                        "name": "include",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Surahs retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/quran.GetAllSurahsSuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/quran.QuranErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/quran.QuranErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new surah in the Quran",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Surahs"
                ],
                "summary": "Create a new surah",
                "parameters": [
                    {
                        "description": "Create surah request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateSurahRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Surah created successfully",
                        "schema": {
                            "$ref": "#/definitions/quran.CreateSurahSuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/quran.QuranErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/quran.QuranErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "core.ErrorDetail": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "reason": {
                    "type": "string"
                }
            }
        },
...
        "dto.BookmarkAyahResponse": {
            "type": "object",
            "properties": {
                "ayahId": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "createdBy": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "updatedBy": {
                    "type": "string"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
...
        "dto.CreateBookmarkAyahRequest": {
            "type": "object",
            "required": [
                "ayahId",
                "createdBy",
                "userId"
            ],
            "properties": {
                "ayahId": {
                    "type": "string"
                },
                "createdBy": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 1
                },
                "userId": {
                    "type": "string"
                }
            }
        },
...
        "dto.CreateLastReadRequest": {
            "type": "object",
            "required": [
                "ayahId",
                "ayahNumber",
                "createdBy",
                "progressPct",
                "surahId",
                "userId"
            ],
            "properties": {
                "ayahId": {
                    "type": "string"
                },
                "ayahNumber": {
                    "type": "integer",
                    "minimum": 1
                },
                "createdBy": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 1
                },
                "progressPct": {
                    "type": "number",
                    "maximum": 100,
                    "minimum": 0
                },
                "surahId": {
                    "type": "string"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
        "dto.CreateProgressHatamRequest": {
            "type": "object",
            "required": [
                "createdBy",
                "juzId",
                "progressPct",
                "startAyahId",
                "userId"
            ],
            "properties": {
                "createdBy": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 1
                },
                "juzId": {
                    "type": "string"
                },
                "progressPct": {
                    "type": "number",
                    "maximum": 100,
                    "minimum": 0
                },
                "startAyahId": {
                    "type": "string"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
...
        "dto.CreateSurahRequest": {
            "type": "object",
            "required": [
                "ayahCount",
                "createdBy",
                "nameArabic",
                "nameEnglish",
                "revelationOrder",
                "revelationPlace"
            ],
            "properties": {
                "ayahCount": {
                    "type": "integer",
                    "minimum": 1
                },
                "createdBy": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 1
                },
                "nameArabic": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 1
                },
                "nameEnglish": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 1
                },
                "revelationOrder": {
                    "type": "integer",
                    "maximum": 114,
                    "minimum": 1
                },
                "revelationPlace": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 1
                }
            }
        },
...
        "quran.BookmarkAyahSuccessResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/dto.BookmarkAyahResponse"
                },
                "message": {
                    "type": "string",
                    "example": "Bookmark retrieved successfully"
                },
                "responseTime": {
                    "type": "string",
                    "example": "15.234ms"
                },
                "status_code": {
                    "type": "integer",
                    "example": 200
                },
                "success": {
                    "type": "boolean",
                    "example": true
                },
                "timestamp": {
                    "type": "string",
                    "example": "2024-01-01T00:00:00Z"
                }
            }
        },
...
        "quran.CreateSurahSuccessResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/dto.SurahResponse"
                },
                "message": {
                    "type": "string",
                    "example": "Surah created successfully"
                },
                "responseTime": {
                    "type": "string",
                    "example": "15.234ms"
                },
                "status_code": {
                    "type": "integer",
                    "example": 201
                },
                "success": {
                    "type": "boolean",
                    "example": true
                },
                "timestamp": {
                    "type": "string",
                    "example": "2024-01-01T00:00:00Z"
                }
            }
        },
...
        "quran.DeleteBookmarkAyahSuccessResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "string",
                    "example": "null"
                },
                "message": {
                    "type": "string",
                    "example": "Bookmark deleted successfully"
                },
                "responseTime": {
                    "type": "string",
                    "example": "15.234ms"
                },
                "status_code": {
                    "type": "integer",
                    "example": 200
                },
                "success": {
                    "type": "boolean",
                    "example": true
                },
                "timestamp": {
                    "type": "string",
                    "example": "2024-01-01T00:00:00Z"
                }
            }
        },
...
        "quran.GetAllSurahsSuccessResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/dto.SurahListResponse"
                },
                "message": {
                    "type": "string",
                    "example": "Surahs retrieved successfully"
                },
                "responseTime": {
                    "type": "string",
                    "example": "15.234ms"
                },
                "status_code": {
                    "type": "integer",
                    "example": 200
                },
                "success": {
                    "type": "boolean",
                    "example": true
                },
                "timestamp": {
                    "type": "string",
                    "example": "2024-01-01T00:00:00Z"
                }
            }
        },
...
        "quran.LastReadSuccessResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/dto.LastReadResponse"
                },
                "message": {
                    "type": "string",
                    "example": "Last read retrieved successfully"
                },
                "responseTime": {
                    "type": "string",
                    "example": "15.234ms"
                },
                "status_code": {
                    "type": "integer",
                    "example": 200
                },
                "success": {
                    "type": "boolean",
                    "example": true
                },
                "timestamp": {
                    "type": "string",
                    "example": "2024-01-01T00:00:00Z"
                }
            }
        },
...
        "quran.ProgressHatamSuccessResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/dto.ProgressHatamResponse"
                },
                "message": {
                    "type": "string",
                    "example": "Progress hatam retrieved successfully"
                },
                "responseTime": {
                    "type": "string",
                    "example": "15.234ms"
                },
                "status_code": {
                    "type": "integer",
                    "example": 200
                },
                "success": {
                    "type": "boolean",
                    "example": true
                },
                "timestamp": {
                    "type": "string",
                    "example": "2024-01-01T00:00:00Z"
                }
            }
        },
        "quran.QuranErrorResponse": {
            "type": "object",
            "properties": {
                "errors": {
                    "$ref": "#/definitions/core.ErrorDetail"
                },
                "message": {
                    "type": "string",
                    "example": "Failed to process request"
                },
                "responseTime": {
                    "type": "string",
                    "example": "15.234ms"
                },
                "status_code": {
                    "type": "integer",
                    "example": 500
                },
                "success": {
                    "type": "boolean",
                    "example": false
                },
                "timestamp": {
                    "type": "string",
                    "example": "2024-01-01T00:00:00Z"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Enter the token with the `Bearer: ` prefix, e.g. \"Bearer abcde12345\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "BasicAuth": {
            "type": "basic"
        }
    },
...
}